New feature: "
TL_lecture-slide-builder

Commit this repo.  V1.

This is great!  New feature, OCR to get the text and name the files with the existing file name but appended with an appropriate title based on text (need to send OCR’d text to Deepseek with prompt “come up with title under a certain number of characters, with no spaces but underscores). Note, this should be a separate process but can run in parallel so as to not interupt the original slide building process

"